a whole bunch of sick shit.

+++ utilslib/src/cmn/utilslib/math/tuple/api/Tup4dBase.java
@@ -5,7 +5,7 @@ public interface Tup4dBase
 	/** Returns the count of Dimensions of the Vector. */
 	default int getDimensions() { return 4; }
 	
-	float get(int index);
+	double get(int index);
 	
 	Tup4d clone();
 }
