+++ utilslib/src/cmn/utilslib/math/tuple/api/Tup6dBase.java
@@ -5,7 +5,7 @@ public interface Tup6dBase
 	/** Returns the count of Dimensions of the Vector. */
 	default int getDimensions() { return 6; }
 	
-	float get(int index);
+	double get(int index);
 	
 	Tup6f clone();
 }
